pyproject.toml
[tool.poetry]
name = "campaign-network-graphs"
version = "0.1.0"
description = ""
authors = ["nicolai <nicolaibthomsen@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
streamlit = "^1.11.0"
pandas = "^1.4.3"
pre-commit = "^2.20.0"
plotly = "^5.9.0"
networkx = "^2.8.5"
numpy = "1.22.4"
google-api-python-client = "^2.54.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.2"
d3graph = "^2.1.1"
PyDrive = "^1.3.1"

[tool.poetry.dev-dependencies]
ipykernel = "^6.15.1"
seaborn = "^0.11.2"
markdownify = "^0.11.2"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
